+++ src/com/esotericsoftware/yamlbeans/YamlWriter.java
@@ -34,8 +34,8 @@ import java.util.Set;
 import com.esotericsoftware.yamlbeans.Beans.Property;
 import com.esotericsoftware.yamlbeans.YamlConfig.WriteClassName;
 import com.esotericsoftware.yamlbeans.YamlConfig.WriteConfig;
-import com.esotericsoftware.yamlbeans.docs.YamlDocument;
-import com.esotericsoftware.yamlbeans.docs.YamlElement;
+import com.esotericsoftware.yamlbeans.document.YamlDocument;
+import com.esotericsoftware.yamlbeans.document.YamlElement;
 import com.esotericsoftware.yamlbeans.emitter.Emitter;
 import com.esotericsoftware.yamlbeans.emitter.EmitterException;
 import com.esotericsoftware.yamlbeans.parser.AliasEvent;
